Supporting TIN.shp, more than 3w Polygons were loaded.
